Remove and relocate Starter Code files
