+++ frontend/starter_code_reference/APIClients/EntityAPIClient.ts
@@ -1,6 +1,6 @@
-import baseAPIClient from "./BaseAPIClient";
-import AUTHENTICATED_USER_KEY from "../constants/AuthConstants";
-import { getLocalStorageObjProperty } from "../utils/LocalStorageUtils";
+import baseAPIClient from "../../src/APIClients/BaseAPIClient";
+import AUTHENTICATED_USER_KEY from "../../src/constants/AuthConstants";
+import { getLocalStorageObjProperty } from "../../src/utils/LocalStorageUtils";
 
 enum EnumField {
   "A",
