+++ frontend/starter_code_reference/HooksDemo/EmojiSection.tsx
