+++ frontend/starter_code_reference/hooks/useFileUploadManager.ts
