+++ frontend/starter_code_reference/types/SampleContextTypes.ts
