+++ frontend/starter_code_reference/utils/CSVUtils.test.ts
@@ -1,4 +1,4 @@
-import * as CSVUtils from "../CSVUtils";
+import * as CSVUtils from "./CSVUtils";
 
 type Pet = {
   name: string;
